requirements.txt
openai
python-dotenv
textblob
matplotlib
nltk